feat: day2
It took more time to parse the data than to solve a problem.
Part 2 requirese to generate all possible variatons of 'array' so I have
to once again use unnest + generate_series and then do some shit with it
or write setup_part2 and not split array into columns or something. Cba